[gocui] Make search highlighting take precedence over selection highlighting
Previously, jumping to the next search result with 'n' would not show the search
result in the selected line. This is especially annoying when there is more than
one search result on a single line, because pressing 'n' seemingly does nothing.

Now we highlight the search result on top of the selection highlight, which
solves the problem nicely.