Fix showing the "YOU ARE HERE" marker for old-style non-interactive rebases
When stopping in an old-style non-interactive rebase (see previous commit for an
explanation of what that means), the YOU ARE HERE marker should point to the
first non-rebasing commit, but it was pointing at the first rebasing commit
instead.

To keep the tests working, we need to set all commits that have an Action to
StatusRebasing, which matches what happens in the real code.