Fix container detection
Running WSL without a container would be treated as native linux and attempt to run `xdg-open` due to `isContainer()` always returning true.

As far as I can tell, the 'container' env var check is not used by lazygit nor any popular containers and is therefore only a way to manually tell lazygit that the process is running in a container.
This check has been kept but updated into uppercase to follow the POSIX standard, which makes this a breaking change.

Fixes #2757. bug introduced in 4d78d76