[gocui] Support searching the model instead of the view
Add a method View.SetModelSearchFunc; if set, it will be used to look for search
results instead of searching the view lines.

This will be needed when we optimize some list views to render only the visible
portion of the view instead of all lines of the list. However, it also has the
benefit that it allows searching for the full text of something that is only
rendered in an abbreviated form in the view; the most notable example being
commit hashes.

This is opt-in, so we still support searching the view as before. This is needed
for two reasons:
- adding a model search func can be a bit of work, so we should only do it when
  there's a benefit
- some views don't even have a model