DROPME Use refreshViewportOnChange and renderOnlyVisibleLines independently
This is a commit only for testing; we'll drop it again before merging.

We now have two separate flags, refreshViewportOnChange and
renderOnlyVisibleLines, but we only have views that either use them both, or
neither. To test whether the logic around them is really independent, change two
other views to use them independently:

- the local branches view is changed to refresh viewport on change, and in order
  to make that testable, we change the branch render function so that the
  selected branch is rendered yellow. This is comparable to the graph node of
  the selected commit being rendered white.
- the remote branches view is changed to render only the visible lines.
  Conveniently, this is a view that uses filtering and not searching, so we
  don't have to bother implementing a model search func.